+++ testutil/key.go
@@ -1,34 +1,30 @@
 package testutil
 
 import (
-	"errors"
+	"fmt"
 
-	"github.com/cosmos/cosmos-sdk/codec"
 	"github.com/cosmos/cosmos-sdk/crypto/keyring"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 )
 
-// GenerateCoinKey generates a new key mnemonic along with its address.
-func GenerateCoinKey(algo keyring.SignatureAlgo, cdc codec.Codec) (sdk.AccAddress, string, error) {
+// GenerateCoinKey generates a new key mnemonic along with its addrress.
+func GenerateCoinKey(algo keyring.SignatureAlgo) (sdk.AccAddress, string, error) {
 	// generate a private key, with mnemonic
-	info, secret, err := keyring.NewInMemory(cdc).NewMnemonic(
+	info, secret, err := keyring.NewInMemory().NewMnemonic(
 		"name",
 		keyring.English,
-		sdk.GetFullBIP44Path(),
+		sdk.GetConfig().GetFullBIP44Path(),
 		keyring.DefaultBIP39Passphrase,
 		algo,
 	)
 	if err != nil {
 		return sdk.AccAddress{}, "", err
 	}
-	addr, err := info.GetAddress()
-	if err != nil {
-		return sdk.AccAddress{}, "", err
-	}
-	return addr, secret, nil
+
+	return sdk.AccAddress(info.GetPubKey().Address()), secret, nil
 }
 
-// GenerateSaveCoinKey generates a new key mnemonic with its address.
+// GenerateSaveCoinKey generates a new key mnemonic with its addrress.
 // If mnemonic is provided then it's used for key generation.
 // The key is saved in the keyring. The function returns error if overwrite=true and the key
 // already exists.
@@ -37,7 +33,6 @@ func GenerateSaveCoinKey(
 	keyName, mnemonic string,
 	overwrite bool,
 	algo keyring.SignatureAlgo,
-	hdPath string,
 ) (sdk.AccAddress, string, error) {
 	exists := false
 	_, err := keybase.Key(keyName)
@@ -47,34 +42,41 @@ func GenerateSaveCoinKey(
 
 	// ensure no overwrite
 	if !overwrite && exists {
-		return sdk.AccAddress{}, "", errors.New("key already exists, overwrite is disabled")
+		return sdk.AccAddress{}, "", fmt.Errorf("key already exists, overwrite is disabled")
 	}
 
 	if exists {
 		if err := keybase.Delete(keyName); err != nil {
-			return sdk.AccAddress{}, "", errors.New("failed to overwrite key")
+			return sdk.AccAddress{}, "", fmt.Errorf("failed to overwrite key")
 		}
 	}
 
 	var (
-		record *keyring.Record
+		info   keyring.Info
 		secret string
 	)
 
-	// generate or recover a new account
 	if mnemonic != "" {
 		secret = mnemonic
-		record, err = keybase.NewAccount(keyName, mnemonic, keyring.DefaultBIP39Passphrase, hdPath, algo)
+		info, err = keybase.NewAccount(
+			keyName,
+			mnemonic,
+			keyring.DefaultBIP39Passphrase,
+			sdk.GetConfig().GetFullBIP44Path(),
+			algo,
+		)
 	} else {
-		record, secret, err = keybase.NewMnemonic(keyName, keyring.English, hdPath, keyring.DefaultBIP39Passphrase, algo)
+		info, secret, err = keybase.NewMnemonic(
+			keyName,
+			keyring.English,
+			sdk.GetConfig().GetFullBIP44Path(),
+			keyring.DefaultBIP39Passphrase,
+			algo,
+		)
 	}
 	if err != nil {
 		return sdk.AccAddress{}, "", err
 	}
 
-	addr, err := record.GetAddress()
-	if err != nil {
-		return nil, "", err
-	}
-	return addr, secret, nil
+	return sdk.AccAddress(info.GetAddress()), secret, nil
 }